Refactor service and chat components into reusable hooks
