+++ src/components/AssistantChat.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import React, { useEffect, useRef } from "react";
 import {
   ActivityIndicator,
   KeyboardAvoidingView,
@@ -12,44 +12,8 @@ import {
 } from "react-native";
 import { Ionicons } from "@expo/vector-icons";
 
-import { isOpenAiConfigured, transcribeAudio } from "../lib/openai";
-import { runBookingAgent } from "../lib/bookingAgent";
-import { BARBERS, type Service } from "../lib/domain";
-
-type AssistantChatCopy = {
-  initialMessage: string;
-  apiKeyWarning: string;
-  contextPrefix: string;
-  quickRepliesTitle: string;
-  quickRepliesToggleShow: string;
-  quickRepliesToggleHide: string;
-  quickReplyAccessibility: (suggestion: string) => string;
-  quickReplies: {
-    existingBookings: string;
-    bookService: string;
-    bookSpecificService: (serviceName: string) => string;
-    barberAvailability: (barberName: string) => string;
-  };
-  inputPlaceholder: string;
-  sendAccessibility: string;
-  suggestionsAccessibility: {
-    show: string;
-    hide: string;
-  };
-  voiceButtonAccessibility: {
-    start: string;
-    stop: string;
-  };
-  errors: {
-    generic: string;
-    missingApiKey: string;
-    voiceWebOnly: string;
-    voiceUnsupported: string;
-    voiceStartFailed: string;
-    noAudio: string;
-    processFailed: string;
-  };
-};
+import type { Service } from "../lib/domain";
+import { useAssistantChat, type AssistantChatCopy } from "../hooks/useAssistantChat";
 
 const DEFAULT_COPY: AssistantChatCopy = {
   initialMessage:
@@ -87,11 +51,6 @@ const DEFAULT_COPY: AssistantChatCopy = {
   },
 };
 
-type DisplayMessage = {
-  role: "assistant" | "user";
-  content: string;
-};
-
 type AssistantChatProps = {
   colors: {
     text: string;
@@ -118,296 +77,37 @@ export default function AssistantChat({
   services,
   copy = DEFAULT_COPY,
 }: AssistantChatProps) {
-  const [messages, setMessages] = useState<DisplayMessage[]>([
-    { role: "assistant", content: copy.initialMessage },
-  ]);
-  const [input, setInput] = useState("");
-  const [pending, setPending] = useState(false);
-  const [voiceTranscribing, setVoiceTranscribing] = useState(false);
-  const [isRecording, setIsRecording] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [suggestionsVisible, setSuggestionsVisible] = useState(false);
-
-  useEffect(() => {
-    setMessages((prev) => {
-      if (prev.length === 0) {
-        return [{ role: "assistant", content: copy.initialMessage }];
-      }
-      const next = [...prev];
-      const first = next[0];
-      if (first?.role === "assistant" && first.content !== copy.initialMessage) {
-        next[0] = { role: "assistant", content: copy.initialMessage };
-        return next;
-      }
-      return prev;
-    });
-  }, [copy.initialMessage]);
-
-  const quickReplies = useMemo(() => {
-    const replies = [copy.quickReplies.existingBookings, copy.quickReplies.bookService];
-
-    services.slice(0, 2).forEach((service) => {
-      replies.push(copy.quickReplies.bookSpecificService(service.name));
-    });
-
-    BARBERS.slice(0, 2).forEach((barber) => {
-      replies.push(copy.quickReplies.barberAvailability(barber.name));
-    });
-
-    return replies;
-  }, [copy.quickReplies, services]);
-
   const scrollRef = useRef<ScrollView>(null);
-  const messagesRef = useRef(messages);
-  const mediaRecorderRef = useRef<any>(null);
-  const recordedChunksRef = useRef<any[]>([]);
 
-  useEffect(() => {
-    messagesRef.current = messages;
-  }, [messages]);
-  useEffect(() => {
-    const trimmed = contextSummary.trim();
-    setMessages((prev) => {
-      const contextPrefix = copy.contextPrefix;
-      const existingIndex = prev.findIndex(
-        (msg) => msg.role === "assistant" && msg.content.startsWith(contextPrefix),
-      );
+  const {
+    messages,
+    input,
+    setInput,
+    pending,
+    error,
+    canSend,
+    assistantEnabled,
+    quickReplies,
+    suggestionsVisible,
+    showSuggestions,
+    hideSuggestions,
+    isRecording,
+    voiceTranscribing,
+    voiceButtonDisabled,
+    handleSend,
+    handleQuickReply,
+    handleVoicePress,
+  } = useAssistantChat({
+    systemPrompt,
+    contextSummary,
+    services,
+    copy,
+    onBookingsMutated,
+  });
 
-      if (!trimmed) {
-        if (existingIndex === -1) return prev;
-        const next = [...prev];
-        next.splice(existingIndex, 1);
-        return next;
-      }
-
-      const contextContent = `${contextPrefix}\n${trimmed}`;
-      if (existingIndex !== -1) {
-        if (prev[existingIndex].content === contextContent) {
-          return prev;
-        }
-        const next = [...prev];
-        next[existingIndex] = { role: "assistant", content: contextContent };
-        return next;
-      }
-
-      const next = [...prev];
-      const insertIndex = next.length > 0 ? 1 : 0;
-      next.splice(insertIndex, 0, { role: "assistant", content: contextContent });
-      return next;
-    });
-  }, [contextSummary, copy.contextPrefix]);
-  useEffect(() => {
-    setMessages((prev) => {
-      const filtered = prev.filter(
-        (msg) =>
-          !(msg.role === "assistant" &&
-            (msg.content === copy.apiKeyWarning || msg.content === DEFAULT_COPY.apiKeyWarning)),
-      );
-      if (isOpenAiConfigured) {
-        return filtered;
-      }
-      if (filtered.some((msg) => msg.role === "assistant" && msg.content === copy.apiKeyWarning)) {
-        return filtered;
-      }
-      return [...filtered, { role: "assistant", content: copy.apiKeyWarning }];
-    });
-  }, [copy.apiKeyWarning, isOpenAiConfigured]);
   useEffect(() => {
     scrollRef.current?.scrollToEnd({ animated: true });
   }, [messages]);
-
-  useEffect(() => {
-    return () => {
-      if (mediaRecorderRef.current) {
-        try {
-          mediaRecorderRef.current.stop();
-        } catch (e) {
-          // ignore cleanup errors
-        }
-      }
-    };
-  }, []);
-
-  const voiceSupported = Platform.OS === "web";
-
-  const canSend = useMemo(() => {
-    return Boolean(input.trim()) && !pending && isOpenAiConfigured && !voiceTranscribing;
-  }, [input, pending, voiceTranscribing]);
-
-  const startVoiceRecording = useCallback(async () => {
-    if (!isOpenAiConfigured) {
-      setError(copy.errors.missingApiKey);
-      return;
-    }
-    const globalNavigator: any = Platform.OS === "web" ? (globalThis as any).navigator : null;
-    if (!globalNavigator?.mediaDevices?.getUserMedia) {
-      setError(copy.errors.voiceWebOnly);
-      return;
-    }
-
-    try {
-      const RecorderCtor = (globalThis as any).MediaRecorder;
-      if (typeof RecorderCtor !== "function") {
-        setError(copy.errors.voiceUnsupported);
-        return;
-      }
-
-      const stream = await globalNavigator.mediaDevices.getUserMedia({ audio: true });
-      recordedChunksRef.current = [];
-      const recorder = new RecorderCtor(stream);
-      recorder.ondataavailable = (event) => {
-        if (event.data && event.data.size > 0) {
-          recordedChunksRef.current.push(event.data);
-        }
-      };
-      mediaRecorderRef.current = recorder;
-      recorder.start();
-      setError(null);
-      setIsRecording(true);
-    } catch (e: any) {
-      const message = e?.message ? String(e.message) : copy.errors.voiceStartFailed;
-      setError(message);
-      setIsRecording(false);
-    }
-  }, [copy.errors.missingApiKey, copy.errors.voiceStartFailed, copy.errors.voiceUnsupported, copy.errors.voiceWebOnly, isOpenAiConfigured]);
-
-  const stopVoiceRecording = useCallback(async () => {
-    const recorder = mediaRecorderRef.current;
-    if (!recorder) return null;
-
-    return new Promise<any>((resolve, reject) => {
-      recorder.onstop = () => {
-        try {
-          const chunks = recordedChunksRef.current;
-          const BlobCtor = (globalThis as any).Blob;
-          const blob = chunks.length && typeof BlobCtor === "function"
-            ? new BlobCtor(chunks, { type: recorder.mimeType || "audio/webm" })
-            : null;
-          recorder.stream.getTracks().forEach((track) => track.stop());
-          mediaRecorderRef.current = null;
-          recordedChunksRef.current = [];
-          resolve(blob);
-        } catch (err) {
-          reject(err);
-        }
-      };
-
-      try {
-        recorder.stop();
-      } catch (err) {
-        mediaRecorderRef.current = null;
-        reject(err);
-      }
-    });
-  }, []);
-
-  const sendMessage = useCallback(
-    async (rawText: string) => {
-      const trimmed = rawText.trim();
-      if (!trimmed || pending) return;
-
-      const userMessage: DisplayMessage = { role: "user", content: trimmed };
-      const nextMessages = [...messagesRef.current, userMessage];
-      setMessages(nextMessages);
-      setPending(true);
-      setError(null);
-
-      try {
-        const reply = await runBookingAgent({
-          systemPrompt,
-          contextSummary,
-          conversation: nextMessages,
-          onBookingsMutated,
-          services,
-        });
-        setMessages((prev) => [...prev, { role: "assistant", content: reply }]);
-      } catch (e: any) {
-        const message = e?.message ? String(e.message) : copy.errors.generic;
-        setError(message);
-      } finally {
-        setPending(false);
-      }
-    },
-    [contextSummary, copy.errors.generic, onBookingsMutated, pending, services, systemPrompt],
-  );
-
-  const handleSend = useCallback(() => {
-    const trimmed = input.trim();
-    if (!trimmed) return;
-    setInput("");
-    void sendMessage(trimmed);
-  }, [input, sendMessage]);
-
-  const handleVoicePress = useCallback(async () => {
-    if (pending || voiceTranscribing) return;
-
-    if (isRecording) {
-      setIsRecording(false);
-      setVoiceTranscribing(true);
-      try {
-        const blob = await stopVoiceRecording();
-        if (!blob) {
-          setError(copy.errors.noAudio);
-          return;
-        }
-        const mimeType = blob.type || "audio/webm";
-        const extension = (() => {
-          if (!mimeType) return "webm";
-          if (mimeType.includes("mp4") || mimeType.includes("m4a")) {
-            return "m4a";
-          }
-          if (mimeType.includes("ogg")) {
-            return "ogg";
-          }
-          if (mimeType.includes("wav")) {
-            return "wav";
-          }
-          if (mimeType.includes("mp3")) {
-            return "mp3";
-          }
-          if (mimeType.includes("webm")) {
-            return "webm";
-          }
-          return "webm";
-        })();
-        const transcript = await transcribeAudio({
-          blob,
-          fileName: `voice-message.${extension}`,
-          mimeType,
-        });
-        await sendMessage(transcript);
-      } catch (e: any) {
-        const message = e?.message ? String(e.message) : copy.errors.processFailed;
-        setError(message);
-      } finally {
-        setVoiceTranscribing(false);
-      }
-    } else {
-      await startVoiceRecording();
-    }
-  }, [
-    copy.errors.noAudio,
-    copy.errors.processFailed,
-    isRecording,
-    pending,
-    sendMessage,
-    startVoiceRecording,
-    stopVoiceRecording,
-    voiceTranscribing,
-  ]);
-
-  const voiceButtonDisabled =
-    !isOpenAiConfigured || pending || voiceTranscribing || (!voiceSupported && !isRecording);
-
-  const handleQuickReply = useCallback(
-    (suggestion: string) => {
-      if (!suggestion) return;
-      setInput("");
-      setSuggestionsVisible(false);
-      void sendMessage(suggestion);
-    },
-    [sendMessage],
-  );
 
   return (
     <KeyboardAvoidingView
@@ -449,9 +149,7 @@ export default function AssistantChat({
           )}
         </ScrollView>
 
-        {error ? (
-          <Text style={[styles.errorText, { color: colors.danger }]}>{error}</Text>
-        ) : null}
+        {error ? <Text style={[styles.errorText, { color: colors.danger }]}>{error}</Text> : null}
 
         {quickReplies.length > 0 && suggestionsVisible ? (
           <View
@@ -462,7 +160,7 @@ export default function AssistantChat({
                 {copy.quickRepliesTitle}
               </Text>
               <Pressable
-                onPress={() => setSuggestionsVisible(false)}
+                onPress={hideSuggestions}
                 accessibilityRole="button"
                 accessibilityLabel={copy.suggestionsAccessibility.hide}
                 hitSlop={8}
@@ -475,13 +173,13 @@ export default function AssistantChat({
                 <Pressable
                   key={suggestion}
                   onPress={() => handleQuickReply(suggestion)}
-                  disabled={pending || voiceTranscribing || !isOpenAiConfigured}
+                  disabled={pending || voiceTranscribing || !assistantEnabled}
                   style={[
                     styles.quickReplyCard,
                     {
                       borderColor: colors.border,
                       backgroundColor: colors.surface,
-                      opacity: pending || voiceTranscribing || !isOpenAiConfigured ? 0.5 : 1,
+                      opacity: pending || voiceTranscribing || !assistantEnabled ? 0.5 : 1,
                     },
                   ]}
                   accessibilityRole="button"
@@ -497,7 +195,7 @@ export default function AssistantChat({
           </View>
         ) : quickReplies.length > 0 ? (
           <Pressable
-            onPress={() => setSuggestionsVisible(true)}
+            onPress={showSuggestions}
             style={[styles.quickRepliesToggle, { borderColor: colors.border, backgroundColor: colors.surface }]}
             accessibilityRole="button"
             accessibilityLabel={copy.suggestionsAccessibility.show}
@@ -526,24 +224,20 @@ export default function AssistantChat({
               isRecording ? copy.voiceButtonAccessibility.stop : copy.voiceButtonAccessibility.start
             }
           >
-            {voiceTranscribing ? (
-              <ActivityIndicator size="small" color={isRecording ? colors.surface : colors.subtext} />
-            ) : (
-              <Ionicons
-                name={isRecording ? "stop" : "mic"}
-                size={18}
-                color={isRecording ? colors.surface : colors.subtext}
-              />
-            )}
+            <Ionicons
+              name={isRecording ? "stop-circle" : "mic"}
+              size={18}
+              color={isRecording ? colors.accentFgOn : colors.subtext}
+            />
           </Pressable>
           <TextInput
             value={input}
             onChangeText={setInput}
             placeholder={copy.inputPlaceholder}
             placeholderTextColor={colors.subtext}
-            multiline
             style={[styles.input, { color: colors.text }]}
-            editable={!pending && isOpenAiConfigured && !voiceTranscribing}
+            multiline
+            editable={!pending && assistantEnabled && !voiceTranscribing}
           />
           <Pressable
             onPress={handleSend}
